extract RUnButton component
